requirements.txt
fastapi
uvicorn[standard]
pdf2image
pytesseract
pillow
img2pdf